+++ rotkehlchen/api/v1/encoding.py
@@ -1991,6 +1991,7 @@ class DataImportSchema(Schema):
                 'nexo',
                 'shapeshift-trades',
                 'uphold',
+                'bisq',
             ),
         ),
     )